fix all attack

+++ src/main/java/com/game/itgame/entity/enemy/EnemyRender.java
@@ -16,11 +16,11 @@ public class EnemyRender extends EntityRender implements EntityInterface {
     public SkillTimer attack1;
     public boolean moveLeft = false, moveRight = false;
 
-    public EnemyRender(int x, int y, GraphicsContext ctx) {
-        this.x = ctx.getCanvas().getWidth()/2-15-2* MapRender.mapFrameSize + x*MapRender.mapFrameSize;
-        this.y = ctx.getCanvas().getHeight()/2-15-2*MapRender.mapFrameSize + y*MapRender.mapFrameSize;
-        mapX = x*MapRender.mapFrameSize;
-        mapY = y*MapRender.mapFrameSize;
+    public EnemyRender(double x, double y, GraphicsContext ctx) {
+        this.x = ctx.getCanvas().getWidth() / 2 - 15 - 2 * MapRender.mapFrameSize + x * MapRender.mapFrameSize;
+        this.y = ctx.getCanvas().getHeight() / 2 - 15 - 2 * MapRender.mapFrameSize + y * MapRender.mapFrameSize;
+        mapX = x * MapRender.mapFrameSize;
+        mapY = y * MapRender.mapFrameSize;
         this.ctx = ctx;
     }
